+++ packages/core/dlight/src/index.d.ts
@@ -5,3 +5,4 @@ export function setGlobal(globalObj: any): void
 export function setDocument(customDocument: any): void
 export const $: typeof escape
 export const View: any
+export const Model: any