+++ packages/core/dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/dlight
 
+## 1.0.0-beta.6
+
+### Patch Changes
+
+- feat: add Model encapsulation
+
 ## 1.0.0-beta.5
 
 ### Patch Changes